Fetch cache backend later to provide thread-safety
According to Django documentation
(https://docs.djangoproject.com/en/2.1/topics/cache/#accessing-the-cache),
when you request a cache backend, a different instance of the backend
will be returned for each thread.

The way the handler works exactly can be seen here:
https://github.com/django/django/blob/0914a2003b1ad50f1d641709da86c14826bf063b/django/core/cache/__init__.py#L57

In its current state, the decorator loads the cache at the point it
wraps the function, that is, when the module is loaded, and then re-uses
it when the wrapped functions are called.

This is a problem, because the instance of the cache backend will always
be the same, even if the thread is now different. In practice this
breaks backends that may rely on thread-safety, like a "per-request
(per-thread)" cache that uses local storage.

This patch simply moves the call to fetch the backend into the inner
functions, to ensure it's always looking it up in the current thread.